+++ modules/alloy/src/main/ts/ephox/alloy/behaviour/docking/DockingApis.ts
@@ -1,23 +1,48 @@
-import { Arr, Fun, Thunk } from '@ephox/katamari';
-import { Classes, Css, Position, Scroll, Traverse } from '@ephox/sugar';
+import { Arr, Fun, Option } from '@ephox/katamari';
+import { Classes, Css, Scroll, Traverse } from '@ephox/sugar';
 
 import * as Boxes from '../../alien/Boxes';
 import * as OffsetOrigin from '../../alien/OffsetOrigin';
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import * as DragCoord from '../../api/data/DragCoord';
 import * as Dockables from './Dockables';
-import { DockingConfig, DockingState } from './DockingTypes';
+import { DockingConfig, DockingMode, DockingState } from './DockingTypes';
+
+const addPx = (num: number) => num + 'px';
 
 const morphToStatic = (component: AlloyComponent, config: DockingConfig): void => {
-  Arr.each([ 'left', 'top', 'position' ], (prop) => Css.remove(component.element(), prop));
+  Arr.each([ 'left', 'top', 'bottom', 'position' ], (prop) => Css.remove(component.element(), prop));
   config.onUndocked(component);
 };
 
-const morphToCoord = (component: AlloyComponent, config: DockingConfig, scroll: Position, origin: Position, morph: DragCoord.CoordAdt): void => {
-  const styles = DragCoord.toStyles(morph, scroll, origin);
-  Css.setAll(component.element(), styles);
-  const method = styles.position === 'fixed' ? config.onDocked : config.onUndocked;
-  method(component);
+const morphToAbsolute = (component: AlloyComponent, config: DockingConfig, left: number, top: number): void => {
+  // Calculate the original offsets
+  const elem = component.element();
+  const scroll = Scroll.get(Traverse.owner(elem));
+  const origin = OffsetOrigin.getOrigin(elem);
+
+  // Calculate the new styles
+  const styles = DragCoord.toStyles(DragCoord.absolute(left, top), scroll, origin);
+
+  // Apply the new ones and fire that the element is no longer docked
+  Css.setOptions(elem, styles);
+  config.onUndocked(component);
+};
+
+const morphToFixed = (component: AlloyComponent, config: DockingConfig, left: number, top: number, bottom: number, mode: DockingMode): void => {
+  // Calculate the new styles
+  const isTop = mode === 'top';
+  const styles: Record<string, Option<string>> = {
+    position: Option.some('fixed'),
+    left: Option.some(addPx(left)),
+    right: Option.none(),
+    top: isTop ? Option.some(addPx(top)) : Option.none(),
+    bottom: !isTop ? Option.some(addPx(bottom)) : Option.none(),
+  };
+
+  // Apply the new styles and fire that the element is docked
+  Css.setOptions(component.element(), styles);
+  config.onDocked(component);
 };
 
 const updateVisibility = (component: AlloyComponent, config: DockingConfig, state: DockingState, viewport: Boxes.Bounds, morphToDocked: boolean = false) => {
@@ -46,30 +71,22 @@ const updateVisibility = (component: AlloyComponent, config: DockingConfig, stat
 const refreshInternal = (component: AlloyComponent, config: DockingConfig, state: DockingState) => {
   // Absolute coordinates (considers scroll)
   const viewport = config.lazyViewport(component);
-
-  const elem = component.element();
-  const doc = Traverse.owner(elem);
-  const scroll = Scroll.get(doc);
-  // PERFORMANCE: OffsetOrigin.getOrigin() is a little bit slow for scrolling performance, so
-  // only do the calculations as required and cache the result to avoid multiple calls
-  const lazyOrigin = Thunk.cached(() => OffsetOrigin.getOrigin(elem));
-
   // If docked then check if we need to hide/show the component
   const isDocked = state.isDocked();
   if (isDocked) {
     updateVisibility(component, config, state, viewport);
   }
 
-  Dockables.getMorph(component, config, viewport, scroll, lazyOrigin).each((morph) => {
+  Dockables.getMorph(component, config, viewport).each((morph) => {
     // Toggle the docked state
     state.setDocked(!isDocked);
     // Apply the morph result
     morph.fold(
       () => morphToStatic(component, config),
-      (x, y) => morphToCoord(component, config, scroll, lazyOrigin(), DragCoord.absolute(x, y)),
-      (x, y) => {
+      (left, top) => morphToAbsolute(component, config, left, top),
+      (left, top, bottom, mode) => {
         updateVisibility(component, config, state, viewport, true);
-        morphToCoord(component, config, scroll, lazyOrigin(), DragCoord.fixed(x, y));
+        morphToFixed(component, config, left, top, bottom, mode);
       },
     );
   });
@@ -82,12 +99,7 @@ const resetInternal = (component: AlloyComponent, config: DockingConfig, state: 
   Dockables.getMorphToOriginal(component, config).each((morph) => {
     morph.fold(
       () => morphToStatic(component, config),
-      (x, y) => {
-        const doc = Traverse.owner(elem);
-        const scroll = Scroll.get(doc);
-        const origin = OffsetOrigin.getOrigin(elem);
-        morphToCoord(component, config, scroll, origin, DragCoord.absolute(x, y));
-      },
+      (left, top) => morphToAbsolute(component, config, left, top),
       Fun.noop
     );
   });